requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
beautifulsoup4==4.13.3
cssselect==1.3.0
Django==5.1.8
django-cors-headers==4.7.0
django-filter==25.1
django-jazzmin==3.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
docx2python==3.5.0
drf-yasg==1.21.10
ecdsa==0.19.1
inflection==0.5.1
lxml==5.3.1
packaging==24.2
paragraphs==1.0.1
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.4.8
pydantic==2.11.2
pydantic_core==2.33.1
PyJWT==2.9.0
PyMySQL==1.1.1
python-decouple==3.8
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
rsa==4.9
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
types-html5lib==1.1.11.20241018
types-lxml==2025.3.30
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
uritemplate==4.1.1
